refactor(pages) : 파일 이름과 폴더구조 변경
