+++ src/pages/SharedPage.module.css
@@ -1,4 +1,7 @@
 .root {
+}
+
+.section {
   display: flex;
   flex-direction: column;
   justify-content: center;
